requirements.txt
streamlit
groq
pyaudio
PyPDF2
pyttsx3
speechrecognition
emoji
python-dotenv
whoosh
googlesearch-python